+++ eslint.config.js
@@ -15,7 +15,7 @@ export default tseslint.config(
     ],
   },
   {
-    files: ['**/*.ts'],
+    files: ['**/*.ts', '**/*.mjs'],
     languageOptions: {
       globals: { ...globals.node },
     },